DineMetra/requirements.txt
# Core Framework
fastapi==0.121.2
uvicorn==0.38.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
alembic==1.12.1

# Data Processing & ETL
pandas==2.2.3
numpy==1.26.4
python-dateutil==2.9.0.post0

# Machine Learning
scikit-learn==1.6.1
scipy==1.15.1
joblib==1.4.2

# Visualization & Monitoring
matplotlib>=3.9.0

# Environment & Configuration
python-dotenv==1.2.1

# HTTP Requests
requests==2.31.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.10.6
pydantic_core==2.27.2
annotated-types==0.7.0

# Utilities
typing_extensions==4.12.2
typing-inspection==0.4.2
click==8.1.7
h11==0.14.0
idna==3.10
sniffio==1.3.1
starlette==0.41.3
anyio==4.2.0
threadpoolctl==3.5.0
pytz==2024.1
tzdata==2024.1
six==1.16.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2